Bucketed trade data

bucket_trades() retrieves open high low close (OHLC) data for the specified symbol/time frame.

bucket_trades(
  binSize = "1m",
  partial = "false",
  symbol = "XBTUSD",
  count = 1000,
  reverse = "true",
  filter = NULL,
  columns = NULL,
  start = NULL,
  startTime = NULL,
  endTime = NULL,
  use_auth = FALSE
)

Arguments

binSize
character string. The time interval to bucket by, must be one of: "1m", "5m", "1h" or "1d".

partial
character string. Either "true" of "false". If "true", will send in-progress (incomplete) bins for the current time period.

symbol
a character string for the instrument symbol. Use available_symbols() to see available symbols.

count
an optional integer to specify the number of rows to return. Maximum of 1000 (the default) per request.

reverse
an optional character string. Either "true" of "false". If "true", result will be ordered with starting with the newest (defaults to "true").

filter
an optional character string for table filtering. Send JSON key/value pairs, such as "{'key':'value'}". See examples.

columns
an optional character vector of column names to return. If NULL, all columns will be returned.

start
an optional integer. Can be used to specify the starting point for results.

startTime
an optional character string. Starting date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

endTime
an optional character string. Ending date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

use_auth
logical. Use TRUE to enable authentication with API key.

Value

bucket_trades() returns a data.frame containing:

timestamp: POSIXct. Date and time of trade.

symbol: character. Instrument ticker.

open: numeric. Opening price for the bucket.

high: numeric. Highest price in the bucket.

low: numeric. Lowest price in the bucket.

close: numeric. Closing price of the bucket.

trades: numeric. Number of trades executed within the bucket.

volume: numeric. Volume in USD.

vwap: numeric. Volume weighted average price.

lastSize: numeric. Size of the last trade executed.

turnover: numeric. How many satoshi were exchanged.

homeNotional: numeric. BTC value of the bucket.

foreignNotional: numeric. USD value of the bucket.

Details

The API will only return 1000 rows per call. If the desired time frame requires more than one API call, consider using map_bucket_trades().

References

urlhttps://www.bitmex.com/api/explorer/#!/Trade/Trade_getBucketed

Examples

# \donttest{

# Return most recent data for symbol `"ETHUSD"` for 1 hour buckets

bucket_trades(
  binSize = "1h",
  symbol = "ETHUSD",
  count = 10
)

#>              timestamp symbol   open   high    low  close trades  volume   vwap
#> 1  2020-05-25 15:00:00 ETHUSD 204.05 204.80 202.55 202.60   3454 3154896 203.81
#> 2  2020-05-25 14:00:00 ETHUSD 203.20 204.50 202.45 204.05   3474 3195656 203.38
#> 3  2020-05-25 13:00:00 ETHUSD 202.60 203.95 202.60 203.20   3107 2856193 203.39
#> 4  2020-05-25 12:00:00 ETHUSD 202.05 202.65 201.50 202.60   2957 2124305 202.21
#> 5  2020-05-25 11:00:00 ETHUSD 202.45 203.90 201.95 202.05   3509 4268361 202.90
#> 6  2020-05-25 10:00:00 ETHUSD 203.10 205.25 201.20 202.45   6941 9943620 203.15
#> 7  2020-05-25 09:00:00 ETHUSD 204.25 204.70 203.10 203.10   2693 2192077 203.93
#> 8  2020-05-25 08:00:00 ETHUSD 203.00 204.50 202.70 204.25   3010 3063474 203.82
#> 9  2020-05-25 07:00:00 ETHUSD 202.20 204.15 201.45 203.00   3833 3535096 203.08
#> 10 2020-05-25 06:00:00 ETHUSD 203.00 203.15 201.75 202.20   1646 1734587 202.34
#>    lastSize     turnover homeNotional foreignNotional
#> 1       200  64299768995     27818.81         5669767
#> 2         2  64992436545     28041.43         5703047
#> 3        50  58092090715     25021.45         5089118
#> 4       500  42955368625     18537.40         3748437
#> 5       450  86602911555     37346.31         7577446
#> 6        35 201996260460     87336.22        17742255
#> 7       100  44702414345     19392.88         3954749
#> 8        25  62438583870     27032.02         5509602
#> 9        25  71790126805     31082.77         6312266
#> 10      250  35096397260     15237.50         3083056

# }