Bucket trade data over an extended period

map_bucket_trades uses purrr::map_dfr under the hood to execute multiple API calls. This is useful when the data you want to return exceeds the maximum 1000 row response limit, but do not want to have to manually call bucket_trades repeatedly.

map_bucket_trades(
  start_date = "2015-09-25 13:00:00",
  end_date = now(tzone = "UTC"),
  binSize = "1d",
  symbol = "XBTUSD",
  partial = "false"
)

Arguments

start_date
The start date of the sample. Defaults to the earliest date available for each available symbol.

end_date
The end date of the sample. Default to today.

binSize
Time interval to bucket by. Must be one of: 1m, 5m, 1h or 1d.

symbol
Instrument symbol. Use available_symbols() to see valid symbols.

partial
If true, will send in-progress (incomplete) bins for the current time period.

Value

`map_bucket_trades` returns a data.frame containing bucketed trade data for the specified time frame.

timestamp

Date and time of trade

symbol

Instrument ticker

open

Opening price for the bucket

high

Highest price in the bucket

low

Lowest price in the bucket

close

Closing price of the bucket

trades

Number of trades executed within the bucket

volume

Volume in USD

vwap

Volume weighted average price

lastSize

Size of the last trade executed

turnover

How many sathoshi were exchanged

homeNotional

BTC value of the bucket

foreignNotional

USD value of the bucket

Details

map_bucket_trades takes a start and end date, and creates a sequence of start dates which are passed in to the startTime parameter in bucket_trades.

The length of time between each start time in each API call is determined by the binSize. For example, "1d" is chosen as the binSize the length of time between start dates will be 1000 days. If "1h" is chosen, it will be 1000 hours etc.

The function will print the number of API calls being sent and provides a progress bar in the console

Public API requests are limited to 30 per minute. Consequently, map_bucket_trades uses purrr::slowly to restrict how often the function is called.

References

API Documentation

Examples

if (FALSE) {
# Get hourly bucketed trade data between 2020-01-01 and 2020-02-01

map_bucket_trades(start_date = "2020-01-01",
                  end_date = "2020-02-01",
                  binSize = "1h")
}