Bucketed trade data (testnet)

tn_bucket_trades() retrieves open high low close (OHLC) data for the specified symbol/time frame.

tn_bucket_trades(
  binSize = "1m",
  partial = "false",
  symbol = "XBTUSD",
  count = 1000,
  reverse = "true",
  filter = NULL,
  columns = NULL,
  start = NULL,
  startTime = NULL,
  endTime = NULL,
  use_auth = FALSE
)

Arguments

binSize
character string. The time interval to bucket by, must be one of: "1m", "5m", "1h" or "1d".

partial
character string. Either "true" of "false". If "true", will send in-progress (incomplete) bins for the current time period.

symbol
a character string for the instrument symbol. Use available_symbols() to see available symbols.

count
an optional integer to specify the number of rows to return. Maximum of 1000 (the default) per request.

reverse
an optional character string. Either "true" of "false". If "true", result will be ordered with starting with the newest (defaults to "true").

filter
an optional character string for table filtering. Send JSON key/value pairs, such as "{'key':'value'}". See examples.

columns
an optional character vector of column names to return. If NULL, all columns will be returned.

start
an optional integer. Can be used to specify the starting point for results.

startTime
an optional character string. Starting date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

endTime
an optional character string. Ending date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

use_auth
logical. Use TRUE to enable authentication with API key.

Details

The API will only return 1000 rows per call. If the desired time frame requires more than one API call, consider using tn_map_bucket_trades().

Examples

# \donttest{

# Return most recent data for symbol `"ETHUSD"` for 1 hour buckets

tn_bucket_trades(
  binSize = "1h",
  symbol = "ETHUSD",
  count = 10
)

#>              timestamp symbol   open   high    low  close trades volume   vwap
#> 1  2020-05-25 15:00:00 ETHUSD 201.95 202.00 201.65 201.95     83   3268 201.95
#> 2  2020-05-25 14:00:00 ETHUSD 201.90 201.95 201.05 201.95     32    888 201.79
#> 3  2020-05-25 13:00:00 ETHUSD 201.95 201.95 201.50 201.90     47   2190 201.79
#> 4  2020-05-25 12:00:00 ETHUSD 202.95 203.00 201.40 201.95     41   2406 201.60
#> 5  2020-05-25 11:00:00 ETHUSD 202.45 203.60 201.05 202.95     43  22867 202.06
#> 6  2020-05-25 10:00:00 ETHUSD 202.75 204.10 201.00 202.45     42  19452 202.07
#> 7  2020-05-25 09:00:00 ETHUSD 203.75 203.75 201.65 202.75     25   2077 202.07
#> 8  2020-05-25 08:00:00 ETHUSD 202.00 203.75 202.00 203.75     28   7776 202.31
#> 9  2020-05-25 07:00:00 ETHUSD 200.50 202.00 200.15 202.00     25    141 201.39
#> 10 2020-05-25 06:00:00 ETHUSD 200.60 201.15 200.15 200.50     42   1639 200.53
#>    lastSize  turnover homeNotional foreignNotional
#> 1         7  65996860    28.811223       5818.3913
#> 2         6  17918085     7.779451       1569.7403
#> 3         5  44190485    19.195246       3873.2751
#> 4        10  48502990    21.024620       4238.3863
#> 5        10 462041020   200.443988      40500.9202
#> 6        10 393063195   170.907062      34535.8795
#> 7         5  41968825    18.351145       3708.1189
#> 8         2 157313090    68.783363      13915.2831
#> 9         1   2839490     1.240730        249.8615
#> 10        6  32865810    14.401924       2887.9283

# }