Cancel all orders (testnet)

Cancel all orders that have been placed using testnet API for a specific symbol, or use a filter to select specific orders.

tn_cancel_all_orders(symbol = NULL, filter = NULL, text = NULL)

Arguments

symbol
string. Optional symbol. If provided, only cancels orders for that symbol.

filter
string. Optional filter for cancellation. Use to only cancel some orders, e.g. '"side": "Buy"'.

text
string. Optional cancellation annotation. e.g. 'Spread Exceeded'.

Value

Returns a data.frame with information about the orders that were cancelled. See https://www.bitmex.com/api/explorer/#!/Order/Order_cancelAll for more information.

Examples

if (FALSE) {
# cancel all "Buy" orders
tn_cancel_all_orders(filter = '{"side": "Buy"}')
}