Bucketed trade data

bucket_trades() retrieves open high low close (OHLC) data for the specified symbol/time frame.

bucket_trades(
  binSize = "1m",
  partial = "false",
  symbol = "XBTUSD",
  count = 1000,
  reverse = "true",
  filter = NULL,
  columns = NULL,
  start = NULL,
  startTime = NULL,
  endTime = NULL,
  use_auth = FALSE
)

Arguments

binSize
character string. The time interval to bucket by, must be one of: "1m", "5m", "1h" or "1d".

partial
character string. Either "true" of "false". If "true", will send in-progress (incomplete) bins for the current time period.

symbol
a character string for the instrument symbol. Use available_symbols() to see available symbols.

count
an optional integer to specify the number of rows to return. Maximum of 1000 (the default) per request.

reverse
an optional character string. Either "true" of "false". If "true", result will be ordered with starting with the newest (defaults to "true").

filter
an optional character string for table filtering. Send JSON key/value pairs, such as "{'key':'value'}". See examples.

columns
an optional character vector of column names to return. If NULL, all columns will be returned.

start
an optional integer. Can be used to specify the starting point for results.

startTime
an optional character string. Starting date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

endTime
an optional character string. Ending date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

use_auth
logical. Use TRUE to enable authentication with API key.

Value

bucket_trades() returns a data.frame containing:

timestamp: POSIXct. Date and time of trade.

symbol: character. Instrument ticker.

open: numeric. Opening price for the bucket.

high: numeric. Highest price in the bucket.

low: numeric. Lowest price in the bucket.

close: numeric. Closing price of the bucket.

trades: numeric. Number of trades executed within the bucket.

volume: numeric. Volume in USD.

vwap: numeric. Volume weighted average price.

lastSize: numeric. Size of the last trade executed.

turnover: numeric. How many satoshi were exchanged.

homeNotional: numeric. BTC value of the bucket.

foreignNotional: numeric. USD value of the bucket.

Details

The API will only return 1000 rows per call. If the desired time frame requires more than one API call, consider using map_bucket_trades().

References

urlhttps://www.bitmex.com/api/explorer/#!/Trade/Trade_getBucketed

Examples

# \donttest{

# Return most recent data for symbol `"ETHUSD"` for 1 hour buckets

bucket_trades(
  binSize = "1h",
  symbol = "ETHUSD",
  count = 10
)

#>              timestamp symbol   open   high    low  close trades  volume   vwap
#> 1  2020-05-24 07:00:00 ETHUSD 209.25 211.35 209.25 210.55   5053 6200269 210.69
#> 2  2020-05-24 06:00:00 ETHUSD 208.65 209.85 208.30 209.25   2616 2702398 209.32
#> 3  2020-05-24 05:00:00 ETHUSD 208.95 209.65 208.65 208.65   2195 2067429 209.29
#> 4  2020-05-24 04:00:00 ETHUSD 209.15 209.55 208.00 208.95   2678 3013380 208.84
#> 5  2020-05-24 03:00:00 ETHUSD 208.00 209.90 208.05 209.15   3131 3373572 209.19
#> 6  2020-05-24 02:00:00 ETHUSD 207.55 208.55 207.40 208.00   1725 2668712 208.21
#> 7  2020-05-24 01:00:00 ETHUSD 206.85 207.95 206.25 207.55   2498 1792194 207.04
#> 8  2020-05-24 00:00:00 ETHUSD 208.20 208.55 206.65 206.85   2256 2185701 207.20
#> 9  2020-05-23 23:00:00 ETHUSD 208.20 208.55 207.65 208.20   1896 1849831 208.18
#> 10 2020-05-23 22:00:00 ETHUSD 208.15 208.30 207.65 208.20   1390 1502366 207.96
#>    lastSize     turnover homeNotional foreignNotional
#> 1         1 130628268680     57467.19        12107297
#> 2      1500  56565457430     24886.64         5209174
#> 3      2300  43269068275     19037.38         3984321
#> 4       150  62928815250     27727.03         5790279
#> 5       248  70571437775     31063.67         6498198
#> 6        10  55564062720     24546.11         5110638
#> 7      5000  37105283725     16437.62         3403230
#> 8       250  45286308905     20081.59         4160797
#> 9       104  38509159340     17072.71         3554141
#> 10        1  31242191550     13867.12         2883713

# }