Bucket trade data over an extended period

map_bucket_trades uses purrr::map_dfr under the hood to execute multiple API calls. This is useful when the data you want to return exceeds the maximum 1000 row response limit, but do not want to have to manually call bucket_trades repeatedly.

map_bucket_trades(
  start_date = "2015-09-25 13:00:00",
  end_date = now(tzone = "UTC"),
  binSize = "1d",
  symbol = "XBT",
  partial = "false"
)

Arguments

start_date
The start date of the sample. Default to the earliest date for api calls

end_date
The end date of the sample. Default to today

binSize
Time interval to bucket by. Must be one of: 1m, 5m, 1h or 1d.

symbol
Instrument symbol. You can also send a timeframe, e.g. XBT:quarterly

partial
If true, will send in-progress (incomplete) bins for the current time period.

Value

`map_bucket_trades` returns a data.frame containing bucketed trade data for the specified time frame.

Details

map_bucket_trades takes a start and end date, and creates a sequence of start dates which are passed in to the startTime parameter in bucket_trades.

The length of time between each start time in each API call is determined by the binSize. For example, "1d" is chosen as the binSize the length of time between start dates will be 1000 days. If "1h" is chosen, it will be 1000 hours etc.

The function will print the number of API calls being sent and provides a progress bar in the consol

Public API requests are limited to 30 per minute. Consequently, the function uses purrr::slowly to ensure this limit is never reached while sending multiple API requests.

References

API Documentation

Examples

# Get hourly bucketed trade data between 2020-01-01 and 2020-02-01

map_bucket_trades(start_date = "2020-01-01", end_date = "2020-02-01")

#> 1 API requests generated.

#> Current limit = 30 requests per minute

#>                    timestamp symbol    open    high     low   close  trades
#> 1   2020-01-01T00:00:00.000Z XBTUSD  7218.5  7298.5  7115.0  7163.0  352988
#> 2   2020-01-02T00:00:00.000Z XBTUSD  7163.0  7234.5  7153.0  7179.0  274470
#> 3   2020-01-03T00:00:00.000Z XBTUSD  7179.0  7183.0  6905.5  6942.0  425292
#> 4   2020-01-04T00:00:00.000Z XBTUSD  6942.0  7412.0  6855.0  7335.0  751545
#> 5   2020-01-05T00:00:00.000Z XBTUSD  7335.0  7399.0  7258.5  7345.0  368114
#> 6   2020-01-06T00:00:00.000Z XBTUSD  7345.0  7494.0  7310.0  7350.5  424155
#> 7   2020-01-07T00:00:00.000Z XBTUSD  7350.5  7799.0  7341.5  7756.5  609959
#> 8   2020-01-08T00:00:00.000Z XBTUSD  7756.5  8214.5  7740.5  8155.5  854480
#> 9   2020-01-09T00:00:00.000Z XBTUSD  8155.5  8462.5  7880.0  8046.0 1012985
#> 10  2020-01-10T00:00:00.000Z XBTUSD  8046.0  8046.0  7740.5  7810.5  681100
#> 11  2020-01-11T00:00:00.000Z XBTUSD  7810.5  8204.0  7682.0  8204.0  925221
#> 12  2020-01-12T00:00:00.000Z XBTUSD  8204.0  8295.0  8006.5  8023.5  679745
#> 13  2020-01-13T00:00:00.000Z XBTUSD  8023.5  8196.0  7955.0  8187.0  498913
#> 14  2020-01-14T00:00:00.000Z XBTUSD  8187.0  8200.0  8033.5  8104.5  409746
#> 15  2020-01-15T00:00:00.000Z XBTUSD  8104.5  8900.0  8104.0  8835.5 1104320
#> 16  2020-01-16T00:00:00.000Z XBTUSD  8835.5  8916.5  8578.0  8816.5  768339
#> 17  2020-01-17T00:00:00.000Z XBTUSD  8816.5  8854.0  8585.5  8719.0  619724
#> 18  2020-01-18T00:00:00.000Z XBTUSD  8719.0  9021.0  8670.5  8901.0  794066
#> 19  2020-01-19T00:00:00.000Z XBTUSD  8901.0  8984.0  8810.0  8916.5  504752
#> 20  2020-01-20T00:00:00.000Z XBTUSD  8916.5  9190.0  8430.0  8700.5  783253
#> 21  2020-01-21T00:00:00.000Z XBTUSD  8700.5  8739.0  8516.0  8632.5  505889
#> 22  2020-01-22T00:00:00.000Z XBTUSD  8632.5  8772.5  8455.0  8727.5  466583
#> 23  2020-01-23T00:00:00.000Z XBTUSD  8727.5  8796.5  8552.0  8664.5  368683
#> 24  2020-01-24T00:00:00.000Z XBTUSD  8664.5  8669.5  8283.5  8392.0  671484
#> 25  2020-01-25T00:00:00.000Z XBTUSD  8392.0  8515.0  8223.5  8425.5  576726
#> 26  2020-01-26T00:00:00.000Z XBTUSD  8425.5  8427.5  8251.0  8326.0  397655
#> 27  2020-01-27T00:00:00.000Z XBTUSD  8326.0  8599.5  8286.0  8597.5  452126
#> 28  2020-01-28T00:00:00.000Z XBTUSD  8597.5  8994.5  8566.0  8900.0  634879
#> 29  2020-01-29T00:00:00.000Z XBTUSD  8900.0  9434.0  8886.0  9400.0  892413
#> 30  2020-01-30T00:00:00.000Z XBTUSD  9400.0  9440.0  9235.0  9294.5  633542
#> 31  2020-01-31T00:00:00.000Z XBTUSD  9294.5  9579.0  9052.0  9510.0  722011
#> 32  2020-02-01T00:00:00.000Z XBTUSD  9510.0  9528.0  9205.0  9339.5  540931
#> 33  2020-02-02T00:00:00.000Z XBTUSD  9339.5  9464.0  9287.5  9382.0  416480
#> 34  2020-02-03T00:00:00.000Z XBTUSD  9382.0  9483.5  9113.0  9325.0  570938
#> 35  2020-02-04T00:00:00.000Z XBTUSD  9325.0  9619.0  9220.0  9282.0  583506
#> 36  2020-02-05T00:00:00.000Z XBTUSD  9282.0  9348.0  9080.0  9166.0  562576
#> 37  2020-02-06T00:00:00.000Z XBTUSD  9166.0  9779.0  9157.0  9630.0  684599
#> 38  2020-02-07T00:00:00.000Z XBTUSD  9630.0  9869.0  9530.0  9763.0  675691
#> 39  2020-02-08T00:00:00.000Z XBTUSD  9763.0  9890.0  9713.0  9815.0  501179
#> 40  2020-02-09T00:00:00.000Z XBTUSD  9815.0  9957.5  9665.5  9908.0  517274
#> 41  2020-02-10T00:00:00.000Z XBTUSD  9908.0 10185.5  9899.0 10172.5  558032
#> 42  2020-02-11T00:00:00.000Z XBTUSD 10172.5 10199.0  9715.5  9862.5  701152
#> 43  2020-02-12T00:00:00.000Z XBTUSD  9862.5 10400.0  9725.0 10286.0  680240
#> 44  2020-02-13T00:00:00.000Z XBTUSD 10286.0 10500.0 10257.0 10363.0  622545
#> 45  2020-02-14T00:00:00.000Z XBTUSD 10363.0 10550.0 10075.0 10241.0  853637
#> 46  2020-02-15T00:00:00.000Z XBTUSD 10241.0 10393.0 10119.0 10369.5  519420
#> 47  2020-02-16T00:00:00.000Z XBTUSD 10369.5 10402.5  9731.0  9907.0  546848
#> 48  2020-02-17T00:00:00.000Z XBTUSD  9907.0 10049.0  9642.0  9934.0  607092
#> 49  2020-02-18T00:00:00.000Z XBTUSD  9934.0  9970.0  9485.0  9715.0  745141
#> 50  2020-02-19T00:00:00.000Z XBTUSD  9715.0 10300.0  9621.0 10195.5  742140
#> 51  2020-02-20T00:00:00.000Z XBTUSD 10195.5 10331.0  9260.5  9596.5  749976
#> 52  2020-02-21T00:00:00.000Z XBTUSD  9596.5  9688.0  9354.0  9607.5  565110
#> 53  2020-02-22T00:00:00.000Z XBTUSD  9607.5  9788.0  9506.0  9694.0  444784
#> 54  2020-02-23T00:00:00.000Z XBTUSD  9694.0  9718.5  9560.5  9675.5  299239
#> 55  2020-02-24T00:00:00.000Z XBTUSD  9675.5 10045.0  9674.5  9966.0  456599
#> 56  2020-02-25T00:00:00.000Z XBTUSD  9966.0 10036.0  9480.0  9660.5  674542
#> 57  2020-02-26T00:00:00.000Z XBTUSD  9660.5  9697.5  9233.5  9317.0  652016
#> 58  2020-02-27T00:00:00.000Z XBTUSD  9317.0  9368.0  8580.0  8780.0  908053
#> 59  2020-02-28T00:00:00.000Z XBTUSD  8780.0  8969.0  8523.0  8820.0  730658
#> 60  2020-02-29T00:00:00.000Z XBTUSD  8820.0  8902.5  8438.0  8717.5  824892
#> 61  2020-03-01T00:00:00.000Z XBTUSD  8717.5  8806.5  8525.5  8531.5  499069
#> 62  2020-03-02T00:00:00.000Z XBTUSD  8531.5  8769.0  8400.0  8532.5  736828
#> 63  2020-03-03T00:00:00.000Z XBTUSD  8532.5  8967.0  8490.0  8918.5  594418
#> 64  2020-03-04T00:00:00.000Z XBTUSD  8918.5  8930.0  8660.0  8757.0  603755
#> 65  2020-03-05T00:00:00.000Z XBTUSD  8757.0  8849.0  8665.0  8756.5  482268
#> 66  2020-03-06T00:00:00.000Z XBTUSD  8756.5  9171.0  8756.0  9070.5  607417
#> 67  2020-03-07T00:00:00.000Z XBTUSD  9070.5  9180.0  9001.0  9158.5  428392
#> 68  2020-03-08T00:00:00.000Z XBTUSD  9158.5  9224.5  8851.0  8900.0  485186
#> 69  2020-03-09T00:00:00.000Z XBTUSD  8900.0  8900.0  8001.0  8038.5  813564
#> 70  2020-03-10T00:00:00.000Z XBTUSD  8038.5  8184.5  7630.5  7928.0 1119012
#> 71  2020-03-11T00:00:00.000Z XBTUSD  7928.0  8150.0  7723.0  7889.5  764935
#> 72  2020-03-12T00:00:00.000Z XBTUSD  7889.5  7985.0  7593.0  7939.0  755014
#> 73  2020-03-13T00:00:00.000Z XBTUSD  7939.0  7967.0  4266.0  4346.0 2038313
#> 74  2020-03-14T00:00:00.000Z XBTUSD  4346.0  5873.5  3596.0  5571.5 2019899
#> 75  2020-03-15T00:00:00.000Z XBTUSD  5571.5  5651.0  5046.0  5160.5 1038656
#> 76  2020-03-16T00:00:00.000Z XBTUSD  5160.5  5974.0  5087.0  5338.5 1058121
#> 77  2020-03-17T00:00:00.000Z XBTUSD  5338.5  5347.5  4394.0  5036.5 1760930
#> 78  2020-03-18T00:00:00.000Z XBTUSD  5036.5  5566.0  4930.0  5327.0 1246806
#> 79  2020-03-19T00:00:00.000Z XBTUSD  5327.0  5455.0  5003.5  5409.5 1122177
#> 80  2020-03-20T00:00:00.000Z XBTUSD  5409.5  6464.0  5260.0  6185.0 1218489
#> 81  2020-03-21T00:00:00.000Z XBTUSD  6185.0  6946.0  5653.0  6204.0 1491646
#> 82  2020-03-22T00:00:00.000Z XBTUSD  6204.0  6448.0  5860.0  6192.0 1045460
#> 83  2020-03-23T00:00:00.000Z XBTUSD  6192.0  6422.5  5737.5  5820.5 1019419
#> 84  2020-03-24T00:00:00.000Z XBTUSD  5820.5  6640.5  5683.0  6486.0 1087715
#> 85  2020-03-25T00:00:00.000Z XBTUSD  6486.0  6876.5  6400.0  6765.0  974631
#> 86  2020-03-26T00:00:00.000Z XBTUSD  6765.0  6984.0  6454.0  6685.5  918092
#> 87  2020-03-27T00:00:00.000Z XBTUSD  6685.5  6793.5  6490.0  6760.5  659431
#> 88  2020-03-28T00:00:00.000Z XBTUSD  6760.5  6875.0  6250.5  6363.5  615327
#> 89  2020-03-29T00:00:00.000Z XBTUSD  6363.5  6364.0  6031.5  6250.0  705022
#> 90  2020-03-30T00:00:00.000Z XBTUSD  6250.0  6277.0  5860.5  5869.0  593828
#> 91  2020-03-31T00:00:00.000Z XBTUSD  5869.0  6647.0  5846.0  6408.0  797887
#> 92  2020-04-01T00:00:00.000Z XBTUSD  6408.0  6524.5  6325.0  6425.0  593295
#> 93  2020-04-02T00:00:00.000Z XBTUSD  6425.0  6723.5  6141.5  6671.5  771841
#> 94  2020-04-03T00:00:00.000Z XBTUSD  6671.5  7299.0  6571.0  6795.5  962439
#> 95  2020-04-04T00:00:00.000Z XBTUSD  6795.5  7049.5  6600.0  6738.0  783094
#> 96  2020-04-05T00:00:00.000Z XBTUSD  6738.0  7097.0  6600.0  6875.0  645145
#> 97  2020-04-06T00:00:00.000Z XBTUSD  6875.0  6920.0  6660.5  6775.5  511370
#> 98  2020-04-07T00:00:00.000Z XBTUSD  6775.5  7391.5  6766.0  7338.0  869947
#> 99  2020-04-08T00:00:00.000Z XBTUSD  7338.0  7475.5  7076.0  7193.0  803792
#> 100 2020-04-09T00:00:00.000Z XBTUSD  7193.0  7430.0  7152.0  7362.5  727230
#> 101 2020-04-10T00:00:00.000Z XBTUSD  7362.5  7373.0  7080.0  7289.5  629724
#>         volume      vwap lastSize     turnover homeNotional foreignNotional
#> 1   1097785557  7201.498     1500 1.524475e+13     152447.5      1097785557
#> 2    762775093  7196.315      100 1.060012e+13     106001.2       762775093
#> 3   1471023953  7050.694     1009 2.086450e+13     208645.0      1471023953
#> 4   3050176897  7217.611       45 4.226289e+13     422628.9      3050176897
#> 5   1183967641  7321.716        1 1.617122e+13     161712.2      1183967641
#> 6   1533245110  7422.252     3616 2.065848e+13     206584.8      1533245110
#> 7   2344491871  7564.297        1 3.099587e+13     309958.7      2344491871
#> 8   3903802488  7949.126        1 4.911025e+13     491102.5      3903802488
#> 9   4906482344  8230.453    12000 5.961718e+13     596171.8      4906482344
#> 10  2634163776  7887.679        1 3.339612e+13     333961.2      2634163776
#> 11  3782453314  7932.730     4500 4.768480e+13     476848.0      3782453314
#> 12  2342781516  8124.797     6869 2.883713e+13     288371.3      2342781516
#> 13  1576712742  8111.616     1500 1.943910e+13     194391.0      1576712742
#> 14  1313199511  8103.728     1403 1.620537e+13     162053.7      1313199511
#> 15  5589882784  8588.114       20 6.509030e+13     650903.0      5589882784
#> 16  3523353461  8759.636      100 4.022379e+13     402237.9      3523353461
#> 17  2445274789  8682.817       31 2.816240e+13     281624.0      2445274789
#> 18  3728596348  8877.841     1000 4.199976e+13     419997.6      3728596348
#> 19  1836510291  8899.964    30000 2.063607e+13     206360.7      1836510291
#> 20  3742743016  8818.342      400 4.244501e+13     424450.1      3742743016
#> 21  2016471911  8637.816     4893 2.334605e+13     233460.5      2016471911
#> 22  1797763495  8643.789     9243 2.079987e+13     207998.7      1797763495
#> 23  1313340524  8670.771      250 1.514742e+13     151474.2      1313340524
#> 24  2892720411  8440.956     3200 3.427255e+13     342725.5      2892720411
#> 25  2520283967  8380.122     1400 3.007552e+13     300755.2      2520283967
#> 26  1484100809  8322.930      500 1.783221e+13     178322.1      1484100809
#> 27  1814758342  8449.514      500 2.147769e+13     214776.9      1814758342
#> 28  3131528831  8757.334      100 3.576082e+13     357608.2      3131528831
#> 29  4571974682  9076.881      600 5.037222e+13     503722.2      4571974682
#> 30  3016724002  9337.068       25 3.230984e+13     323098.4      3016724002
#> 31  3871834826  9376.465      333 4.129492e+13     412949.2      3871834826
#> 32  2429182170  9338.812      100 2.601323e+13     260132.3      2429182170
#> 33  1606928169  9376.465       50 1.713826e+13     171382.6      1606928169
#> 34  2558075884  9363.296      100 2.732121e+13     273212.1      2558075884
#> 35  2901398321  9357.163      100 3.100860e+13     310086.0      2901398321
#> 36  2851149240  9200.478    20000 3.099011e+13     309901.1      2851149240
#> 37  3835754054  9476.876    20000 4.047533e+13     404753.3      3835754054
#> 38  3831983024  9717.229       11 3.943634e+13     394363.4      3831983024
#> 39  2479752042  9787.609     5000 2.533641e+13     253364.1      2479752042
#> 40  2628334844  9827.044       31 2.674822e+13     267482.2      2628334844
#> 41  2935217599 10083.695        5 2.910968e+13     291096.8      2935217599
#> 42  3654168022  9923.588        1 3.682431e+13     368243.1      3654168022
#> 43  3819505406 10032.103     4276 3.807592e+13     380759.2      3819505406
#> 44  3195518429 10358.401        1 3.085245e+13     308524.5      3195518429
#> 45  4590116972 10298.661        5 4.457180e+13     445718.0      4590116972
#> 46  2321594602 10261.673       25 2.262491e+13     226249.1      2321594602
#> 47  2793702129 10068.466      898 2.774925e+13     277492.5      2793702129
#> 48  2631092637  9882.399       31 2.662598e+13     266259.8      2631092637
#> 49  3536302919  9698.380      100 3.646635e+13     364663.5      3536302919
#> 50  3722933606  9903.932      230 3.759049e+13     375904.9      3722933606
#> 51  4113270117  9963.136      333 4.128788e+13     412878.8      4113270117
#> 52  2896538588  9574.876      374 3.025396e+13     302539.6      2896538588
#> 53  1943485619  9689.923      198 2.005788e+13     200578.8      1943485619
#> 54  1047504190  9655.306       25 1.084946e+13     108494.6      1047504190
#> 55  2129179142  9901.970      154 2.150363e+13     215036.3      2129179142
#> 56  3127248830  9734.255        1 3.212631e+13     321263.1      3127248830
#> 57  2825728347  9472.388       38 2.983278e+13     298327.8      2825728347
#> 58  4672566740  9004.953     3000 5.189074e+13     518907.4      4672566740
#> 59  3233578879  8776.549    12000 3.684555e+13     368455.5      3233578879
#> 60  3535037527  8649.771       31 4.086951e+13     408695.1      3535037527
#> 61  1624579497  8681.309     3000 1.871358e+13     187135.8      1624579497
#> 62  2865341712  8566.044      200 3.345196e+13     334519.6      2865341712
#> 63  2260357992  8757.334       31 2.581117e+13     258111.7      2260357992
#> 64  2506496796  8787.346     1999 2.852515e+13     285251.5      2506496796
#> 65  1706557363  8756.567      106 1.948936e+13     194893.6      1706557363
#> 66  2680324726  9013.069     1000 2.974020e+13     297402.0      2680324726
#> 67  1771717908  9100.009      118 1.947022e+13     194702.2      1771717908
#> 68  2109118200  9042.409    58534 2.332666e+13     233266.6      2109118200
#> 69  3565316312  8457.375        5 4.215673e+13     421567.3      3565316312
#> 70  4800156501  7856.077     2000 6.110598e+13     611059.8      4800156501
#> 71  3415078897  7951.654     3935 4.294894e+13     429489.4      3415078897
#> 72  2616658378  7816.775        1 3.347698e+13     334769.8      2616658378
#> 73  9879605680  6227.038     2173 1.586567e+14    1586567.2      9879605680
#> 74  8470456293  5061.241     2000 1.673668e+14    1673667.8      8470456293
#> 75  2284697834  5382.711       33 4.244679e+13     424467.9      2284697834
#> 76  2416784951  5391.708      518 4.482461e+13     448246.1      2416784951
#> 77  3767784970  4862.394       75 7.749100e+13     774910.0      3767784970
#> 78  2483878294  5301.103       54 4.685687e+13     468568.7      2483878294
#> 79  2329210649  5263.989       31 4.424888e+13     442488.8      2329210649
#> 80  2961750416  5867.167       31 5.048253e+13     504825.3      2961750416
#> 81  3783695861  6381.214     5000 5.929629e+13     592962.9      3783695861
#> 82  2179889762  6160.670      500 3.538592e+13     353859.2      2179889762
#> 83  2206628792  6075.703     1000 3.632075e+13     363207.5      2206628792
#> 84  2720999038  6171.316       50 4.409199e+13     440919.9      2720999038
#> 85  2242836537  6654.688     1110 3.370470e+13     337047.0      2242836537
#> 86  2234063864  6675.122       31 3.346887e+13     334688.7      2234063864
#> 87  1532859653  6660.450       60 2.301456e+13     230145.6      1532859653
#> 88  1641982373  6648.936     1900 2.469597e+13     246959.7      1641982373
#> 89  1883833066  6208.866      100 3.034115e+13     303411.5      1883833066
#> 90  1502675683  6070.540      662 2.475372e+13     247537.2      1502675683
#> 91  2146883987  6298.419    14859 3.408742e+13     340874.2      2146883987
#> 92  1367303451  6440.809     3600 2.122949e+13     212294.9      1367303451
#> 93  2207635222  6347.191    23472 3.478202e+13     347820.2      2207635222
#> 94  2972605816  6806.425     2000 4.367501e+13     436750.1      2972605816
#> 95  2140975954  6832.935      308 3.133505e+13     313350.5      2140975954
#> 96  1619522774  6796.710        6 2.382950e+13     238295.0      1619522774
#> 97  1148471580  6787.945        2 1.691957e+13     169195.7      1148471580
#> 98  2341814274  7094.714     4445 3.300838e+13     330083.8      2341814274
#> 99  2098135699  7311.545     2000 2.869739e+13     286973.9      2098135699
#> 100 1776328298  7302.468        1 2.432659e+13     243265.9      1776328298
#> 101 1676985467  7278.550       98 2.304100e+13     230410.0      1676985467