Bucketed trade data (testnet)

tn_bucket_trades() retrieves open high low close (OHLC) data for the specified symbol/time frame.

tn_bucket_trades(
  binSize = "1m",
  partial = "false",
  symbol = "XBTUSD",
  count = 1000,
  reverse = "true",
  filter = NULL,
  columns = NULL,
  start = NULL,
  startTime = NULL,
  endTime = NULL,
  use_auth = FALSE
)

Arguments

binSize
character string. The time interval to bucket by, must be one of: "1m", "5m", "1h" or "1d".

partial
character string. Either "true" of "false". If "true", will send in-progress (incomplete) bins for the current time period.

symbol
a character string for the instrument symbol. Use available_symbols() to see available symbols.

count
an optional integer to specify the number of rows to return. Maximum of 1000 (the default) per request.

reverse
an optional character string. Either "true" of "false". If "true", result will be ordered with starting with the newest (defaults to "true").

filter
an optional character string for table filtering. Send JSON key/value pairs, such as "{'key':'value'}". See examples.

columns
an optional character vector of column names to return. If NULL, all columns will be returned.

start
an optional integer. Can be used to specify the starting point for results.

startTime
an optional character string. Starting date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

endTime
an optional character string. Ending date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

use_auth
logical. Use TRUE to enable authentication with API key.

Details

The API will only return 1000 rows per call. If the desired time frame requires more than one API call, consider using tn_map_bucket_trades().

Examples

# \donttest{

# Return most recent data for symbol `"ETHUSD"` for 1 hour buckets

tn_bucket_trades(
  binSize = "1h",
  symbol = "ETHUSD",
  count = 10
)

#>              timestamp symbol   open   high    low  close trades volume   vwap
#> 1  2020-05-23 09:00:00 ETHUSD 209.70 209.45 206.95 207.05     44  11411 207.70
#> 2  2020-05-23 08:00:00 ETHUSD 208.05 209.80 208.05 209.70     19     80 208.94
#> 3  2020-05-23 07:00:00 ETHUSD 208.10 210.00 208.05 208.05     31   1779 208.27
#> 4  2020-05-23 06:00:00 ETHUSD 210.00 210.00 207.70 208.10     45  18122 209.83
#> 5  2020-05-23 05:00:00 ETHUSD 210.40 210.55 209.40 210.00     37   4804 209.84
#> 6  2020-05-23 04:00:00 ETHUSD 210.25 210.60 210.25 210.40     57   3588 210.34
#> 7  2020-05-23 03:00:00 ETHUSD 209.00 210.60 209.20 210.25     98   4166 209.97
#> 8  2020-05-23 02:00:00 ETHUSD 209.00 209.20 209.00 209.00     23     96 209.10
#> 9  2020-05-23 01:00:00 ETHUSD 207.35 209.20 206.70 209.00     52   3495 208.48
#> 10 2020-05-23 00:00:00 ETHUSD 208.65 208.65 206.00 207.35     42   1892 207.63
#>    lastSize  turnover homeNotional foreignNotional
#> 1         6 236995630  104.9388554      21794.8412
#> 2         1   1671485    0.7393601        154.4787
#> 3         5  37051230   16.4341925       3422.7561
#> 4         1 380245755  167.6439009      35175.9884
#> 5         4 100804175   44.5932418       9357.1730
#> 6         1  75468235   33.3230495       7009.0097
#> 7         6  87471095   38.6061763       8105.9249
#> 8         6   2007335    0.8878318        185.6433
#> 9        50  72860925   32.2017795       6713.2437
#> 10        3  39282615   17.3454632       3601.3668

# }