Bucketed trade data (testnet)

tn_bucket_trades() retrieves open high low close (OHLC) data for the specified symbol/time frame.

tn_bucket_trades(
  binSize = "1m",
  partial = "false",
  symbol = "XBTUSD",
  count = 1000,
  reverse = "true",
  filter = NULL,
  columns = NULL,
  start = NULL,
  startTime = NULL,
  endTime = NULL,
  use_auth = FALSE
)

Arguments

binSize
character string. The time interval to bucket by, must be one of: "1m", "5m", "1h" or "1d".

partial
character string. Either "true" of "false". If "true", will send in-progress (incomplete) bins for the current time period.

symbol
a character string for the instrument symbol. Use available_symbols() to see available symbols.

count
an optional integer to specify the number of rows to return. Maximum of 1000 (the default) per request.

reverse
an optional character string. Either "true" of "false". If "true", result will be ordered with starting with the newest (defaults to "true").

filter
an optional character string for table filtering. Send JSON key/value pairs, such as "{'key':'value'}". See examples.

columns
an optional character vector of column names to return. If NULL, all columns will be returned.

start
an optional integer. Can be used to specify the starting point for results.

startTime
an optional character string. Starting date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

endTime
an optional character string. Ending date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

use_auth
logical. Use TRUE to enable authentication with API key.

Details

The API will only return 1000 rows per call. If the desired time frame requires more than one API call, consider using tn_map_bucket_trades().

Examples

# \donttest{

# Return most recent data for symbol `"ETHUSD"` for 1 hour buckets

tn_bucket_trades(
  binSize = "1h",
  symbol = "ETHUSD",
  count = 10
)

#>              timestamp symbol   open   high    low  close trades volume   vwap
#> 1  2020-05-31 09:00:00 ETHUSD 225.45 225.45 222.55 222.55     13     48 224.94
#> 2  2020-05-31 08:00:00 ETHUSD 225.45 225.45 223.15 225.45     24    204 224.03
#> 3  2020-05-31 07:00:00 ETHUSD 225.45 225.45 223.10 225.45     11    148 225.44
#> 4  2020-05-31 06:00:00 ETHUSD 225.45 225.45 222.20 225.45     26    441 223.64
#> 5  2020-05-31 05:00:00 ETHUSD 222.95 225.45 222.20 225.45     37 186241 224.67
#> 6  2020-05-31 04:00:00 ETHUSD 222.50 222.95 222.35 222.95     15  13770 222.51
#> 7  2020-05-31 03:00:00 ETHUSD 232.70 232.70 222.35 222.50     22   1295 222.54
#> 8  2020-05-31 02:00:00 ETHUSD 232.95 232.95 226.05 232.70     17     77 232.74
#> 9  2020-05-31 01:00:00 ETHUSD 232.95 232.95 222.35 232.95     33    677 231.62
#> 10 2020-05-31 00:00:00 ETHUSD 232.95 232.95 232.20 232.95     24  50177 232.95
#>    lastSize   turnover homeNotional foreignNotional
#> 1         5    1079665    0.4580544        103.0305
#> 2         2    4570200    1.9472527        436.2447
#> 3         4    3336425    1.4155013        319.1023
#> 4         2    9862140    4.2212987        944.0144
#> 5        15 4184102930 1786.9799013     401464.1915
#> 6       870  306393605  132.3119386      29440.4559
#> 7         2   28818640   12.4795613       2777.1724
#> 8         4    1792050    0.7436956        173.0830
#> 9         4   15680570    6.5392478       1514.6055
#> 10        9 1168871490  486.2303879     113267.2017

# }