Trade data

trades() retrieves information regarding individual trades that have taken place on the exchange for the given symbol / time frame.

trades(
  symbol = "XBTUSD",
  filter = NULL,
  columns = NULL,
  count = 1000,
  start = NULL,
  reverse = "true",
  startTime = NULL,
  endTime = NULL
)

Arguments

symbol
Instrument symbol. Use available_symbols() to see valid symbols.

filter
Generic table filter. Send JSON key/value pairs, such as "{'key':'value'}".

columns
Array of column names to fetch. If omitted, will return all columns.

count
Number of results to fetch. Maximum of 1000 (the default) per request.

start
Starting point for results.

reverse
If true, will sort results newest first (default = "true").

startTime
Starting date filter for results.

endTime
Ending date filter for results.

Value

trades() returns a data.frame containing information for executed trades for the given arguments.

timestamp

Date and time of trade

symbol

Instrument ticker

side

Whether the trade was buy or sell

size

Size of the trade

price

Price the trade was executed at

tickDirection

Indicates if the trade price was higher, lower or the same as the previous trade price

trdMatchID

Unique trade ID

grossValue

How many sathoshi were exchanged. 1 satosi = 0.00000001 BTC

homeNotional

BTC value of the trade

foreignNotional

USD value of the trade

References

API Documentation

Examples

if (FALSE) {
# Return 1000 most recent trades for symbol "XBTUSD".
trades(symbol = "XBTUSD")

# Use filter for very specific values: Return trade data executed at 12:15.
trades(symbol = "XBTUSD", filter = "{'timestamp.minute':'12:15'}")

# Also possible to combine more than one filter.
trades(symbol = "XBTUSD", filter = "{'timestamp.minute':'12:15', 'size':10000}")
}