Individual trade data

trades() retrieves data regarding individual trades that have been executed on the exchange.

trades(
  symbol = "XBTUSD",
  count = 1000,
  reverse = "true",
  filter = NULL,
  columns = NULL,
  start = NULL,
  startTime = NULL,
  endTime = NULL,
  use_auth = FALSE
)

Arguments

symbol
a character string for the instrument symbol. Use available_symbols() to see available symbols.

count
an optional integer to specify the number of rows to return. Maximum of 1000 (the default) per request.

reverse
an optional character string. Either "true" of "false". If "true", result will be ordered with starting with the newest (defaults to "true").

filter
an optional character string for table filtering. Send JSON key/value pairs, such as "{'key':'value'}". See examples.

columns
an optional character vector of column names to return. If NULL, all columns will be returned.

start
an optional integer. Can be used to specify the starting point for results.

startTime
an optional character string. Starting date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

endTime
an optional character string. Ending date for results in the format "yyyy-mm-dd" or "yyyy-mm-dd hh-mm-ss".

use_auth
logical. Use TRUE to enable authentication with API key.

Value

trades() returns a data.frame containing:

timestamp: POSIXct. Date and time of trade.

symbol: character. The instrument ticker.

side: character. Whether the trade was buy or sell.

size: numeric. Size of the trade.

price: numeric. Price the trade was executed at

tickDirection: character. Indicates if the trade price was higher, lower or the same as the previous trade price.

trdMatchID: character. Unique trade ID.

grossValue: numeric. How many satoshi were exchanged. 1 satoshi = 0.00000001 BTC.

homeNotional: numeric. BTC value of the trade.

foreignNotional: numeric. USD value of the trade.

References

https://www.bitmex.com/api/explorer/#!/Trade/Trade_get

Examples

# \donttest{
# Return 1000 most recent trades for symbol "XBTUSD".
trades(symbol = "XBTUSD", count = 10)

#>              timestamp symbol side  size  price tickDirection
#> 1  2020-05-22 13:18:42 XBTUSD  Buy     4 9146.5      PlusTick
#> 2  2020-05-22 13:18:42 XBTUSD Sell     4 9146.0     MinusTick
#> 3  2020-05-22 13:18:42 XBTUSD  Buy     4 9146.5      PlusTick
#> 4  2020-05-22 13:18:42 XBTUSD Sell     4 9146.0     MinusTick
#> 5  2020-05-22 13:18:41 XBTUSD  Buy     4 9146.5      PlusTick
#> 6  2020-05-22 13:18:41 XBTUSD Sell     4 9146.0     MinusTick
#> 7  2020-05-22 13:18:39 XBTUSD  Buy     4 9146.5      PlusTick
#> 8  2020-05-22 13:18:39 XBTUSD Sell 30000 9146.0     MinusTick
#> 9  2020-05-22 13:18:35 XBTUSD  Buy     3 9146.5  ZeroPlusTick
#> 10 2020-05-22 13:18:34 XBTUSD  Buy   500 9146.5  ZeroPlusTick
#>                              trdMatchID grossValue homeNotional foreignNotional
#> 1  304836b7-0fbc-3b05-83f2-8dbc96cec8a8      43732   0.00043732               4
#> 2  952f72e5-aeb3-ab4d-71c7-918ca96a4963      43736   0.00043736               4
#> 3  3deddd40-8df3-1128-38eb-90236414aaac      43732   0.00043732               4
#> 4  cc8393dd-6c17-df0a-f3fd-75625ed41f38      43736   0.00043736               4
#> 5  1297795d-755b-60ef-0ebc-1a637eb1b2cc      43732   0.00043732               4
#> 6  29b4c73b-1bc0-8604-eb9b-af5c33305476      43736   0.00043736               4
#> 7  dc9ce154-9053-716a-15a8-3a68835bd59b      43732   0.00043732               4
#> 8  515134e9-34a8-062e-95de-03c4c27a312f  328020000   3.28020000           30000
#> 9  c0202508-f0c4-e154-fe33-e16477e92a20      32799   0.00032799               3
#> 10 c6e7c142-df97-443b-58af-8cb43605daee    5466500   0.05466500             500

# Use filter for very specific values: Return trade data executed at 12:15.
trades(
  symbol = "XBTUSD",
  filter = "{'timestamp.minute':'12:15'}",
  count = 10)

#>              timestamp symbol side  size price tickDirection
#> 1  2020-05-22 12:15:59 XBTUSD  Buy  6119  9166  ZeroPlusTick
#> 2  2020-05-22 12:15:59 XBTUSD  Buy 10481  9166  ZeroPlusTick
#> 3  2020-05-22 12:15:58 XBTUSD  Buy    37  9166  ZeroPlusTick
#> 4  2020-05-22 12:15:58 XBTUSD  Buy    23  9166  ZeroPlusTick
#> 5  2020-05-22 12:15:58 XBTUSD  Buy    40  9166  ZeroPlusTick
#> 6  2020-05-22 12:15:57 XBTUSD  Buy  3427  9166  ZeroPlusTick
#> 7  2020-05-22 12:15:57 XBTUSD  Buy    23  9166  ZeroPlusTick
#> 8  2020-05-22 12:15:57 XBTUSD  Buy   323  9166  ZeroPlusTick
#> 9  2020-05-22 12:15:57 XBTUSD  Buy    23  9166  ZeroPlusTick
#> 10 2020-05-22 12:15:57 XBTUSD  Buy   238  9166  ZeroPlusTick
#>                              trdMatchID grossValue homeNotional foreignNotional
#> 1  dea3de5f-de21-4f5e-6276-7b54edd3a60b   66758290    0.6675829            6119
#> 2  325e8f06-8c41-2744-9fdd-b2e7b0cd734e  114347710    1.1434771           10481
#> 3  1c269a62-d9ae-287c-4704-07a436f0e87c     403670    0.0040367              37
#> 4  e15abbee-810d-c092-4697-6bb7956dabbd     250930    0.0025093              23
#> 5  8f68fc34-4701-25be-e8ec-10ebd978d2d4     436400    0.0043640              40
#> 6  cc3b5021-bb8a-44df-dc8c-aee5da524f92   37388570    0.3738857            3427
#> 7  a946edab-f555-391f-f8ed-5d3f1b45c25e     250930    0.0025093              23
#> 8  d2646bc1-4108-cd8a-0176-30ab9428fc84    3523930    0.0352393             323
#> 9  e89fea41-1f6f-f6e6-9b5f-1440f5607d0a     250930    0.0025093              23
#> 10 aab507e3-297a-3021-b594-c11c1d18e26d    2596580    0.0259658             238

# Also possible to combine more than one filter.
trades(
  symbol = "XBTUSD",
  filter = "{'timestamp.minute':'12:15', 'size':10000}",
  count = 10)

#>              timestamp symbol side  size  price tickDirection
#> 1  2020-05-22 12:15:50 XBTUSD  Buy 10000 9166.0  ZeroPlusTick
#> 2  2020-05-22 12:15:49 XBTUSD Sell 10000 9165.5 ZeroMinusTick
#> 3  2020-05-22 12:15:49 XBTUSD  Buy 10000 9165.5  ZeroPlusTick
#> 4  2020-05-22 12:15:49 XBTUSD  Buy 10000 9165.0  ZeroPlusTick
#> 5  2020-05-22 12:15:49 XBTUSD  Buy 10000 9162.5  ZeroPlusTick
#> 6  2020-05-22 12:15:49 XBTUSD  Buy 10000 9161.0  ZeroPlusTick
#> 7  2020-05-22 12:15:49 XBTUSD  Buy 10000 9160.5  ZeroPlusTick
#> 8  2020-05-22 12:15:49 XBTUSD  Buy 10000 9160.0  ZeroPlusTick
#> 9  2020-05-22 12:15:45 XBTUSD  Buy 10000 9155.5  ZeroPlusTick
#> 10 2020-05-22 12:15:45 XBTUSD  Buy 10000 9155.5  ZeroPlusTick
#>                              trdMatchID grossValue homeNotional foreignNotional
#> 1  e9b32768-fb1b-c110-da91-eaa9cf376d5a  109100000       1.0910           10000
#> 2  f976fc35-286e-aa09-2d7d-2eb5518a5135  109100000       1.0910           10000
#> 3  170c0669-432a-43ed-428a-b93b3b4cb814  109100000       1.0910           10000
#> 4  7e533676-96c8-203c-cec9-cfeb779bd149  109110000       1.0911           10000
#> 5  ef7be992-8f4a-4db8-ba88-743a9aaa8ba1  109140000       1.0914           10000
#> 6  8a4b6722-fc9a-1162-7f6b-61dc417c1a2b  109160000       1.0916           10000
#> 7  39b40593-c102-5aa7-d0d3-d996c111684d  109160000       1.0916           10000
#> 8  41c733b3-5f1e-f16a-6aae-efd177855ac7  109170000       1.0917           10000
#> 9  010109f9-61a3-5996-9694-fc46bf9f2c32  109220000       1.0922           10000
#> 10 cd4b82b4-7ad6-4960-f176-c4f3384cb471  109220000       1.0922           10000

# }